requirements.txt
absl-py==2.3.0
accelerate==1.7.0
certifi==2025.4.26
charset-normalizer==3.4.2
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.3.2
cycler==0.12.1
diffusers @ git+https://github.com/huggingface/diffusers.git@73a9d5856f2d7ae3637c484d83cd697284ad3962
filelock==3.18.0
fonttools==4.58.1
fsspec==2025.5.1
ftfy==6.3.1
grpcio==1.72.1
hf-xet==1.1.3
huggingface-hub==0.32.4
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
packaging==25.0
pandas==2.3.0
peft==0.15.2
pillow==11.2.1
protobuf==6.31.1
psutil==7.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
six==1.17.0
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.21.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
transformers==4.52.4
triton==3.3.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
zipp==3.22.0
